FluteKeyboard: add alternatives keys implementation
Added logic to handle alternatives keys. Now inside the layout the user
can specify alternatives keys by extending the key string like 'iíìîï'.

When the layout is parsed the first char is used for the button text and
the remaining are used as alternatives. Alternatives keys are inserted
by long pressing the button and overing the needed alternatives.

Signed-off-by: Andrea Ricchi <andrea.ricchi@amarulasolutions.com>