Propagate theme from FluteKeyboard to TextKey
Themes were not properly propagated, resulting in properties of TextStyle, such as font weight, not being effectively applied in the keyboard's text keys.

This adds a nullable theme member to the classes TextKey, AlphanumericKeyboard and NumericKeyboard and
uses that theme member in proper places instead of using a local instance of FluteKeyboardTheme.

Signed-off-by: Nicolas Gry <nicolas.gry@madic.com>